+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [0.11.9] - not yet released
+
+### Removed
+
+- Deprecated context manager on `Streamer`. Instead, use the object directly.
+
+### Added or enhanced
+
+- `mpservice.streamer.Parmapper.__init__` takes two new arguments `executor_initializer`
+  and `executor_init_args`.
+- Simplifications to the implementation of `streamer.py`, making use of `GeneratorExit`.
+
+
 ## [0.11.8] - 2022-12-21
 
 The two largest efforts of this release are documentation and "streamer" refactor.